# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dodge Club API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  isAdmin: boolean;
  memberSince: string;
  eventsAttended: number;
  medalsEarned: number;
  avatarUrl?: string | null;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  location: string;
  ticketUrl?: string | null;
  imageUrl?: string | null;
  isUpcoming: boolean;
  attendeeCount: number;
}

export interface AttendanceRecord {
  id: number;
  userId: number;
  eventId: number;
  attendedAt: string;
  earnedMedal: boolean;
  event: Event;
}

export interface Achievement {
  id: string;
  title: string;
  description: string;
  icon: string;
  unlockedAt?: string | null;
  unlocked: boolean;
}

export interface CreateEventRequest {
  title: string;
  description: string;
  date: string;
  location: string;
  ticketUrl?: string | null;
  imageUrl?: string | null;
}

export interface Post {
  id: number;
  title: string;
  content: string;
  imageUrl?: string | null;
  createdAt: string;
  authorName: string;
  isMembersOnly: boolean;
}

export interface CreatePostRequest {
  title: string;
  content: string;
  imageUrl?: string | null;
  isMembersOnly: boolean;
}

export interface MerchProduct {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl?: string | null;
  buyUrl?: string | null;
  category: string;
  inStock: boolean;
}

export interface CommunityStats {
  totalEvents: number;
  totalMembers: number;
  totalTicketsSold: number;
  totalMedalsAwarded: number;
}