# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Dodge Club API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: events
    description: Event operations
  - name: attendance
    description: Attendance operations
  - name: posts
    description: Message board operations
  - name: merch
    description: Merchandise operations
  - name: tickets
    description: Ticket/event listing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{userId}/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile with stats
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{userId}/attendance:
    get:
      operationId: getUserAttendance
      tags: [attendance]
      summary: Get user attendance history
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"

  /users/{userId}/achievements:
    get:
      operationId: getUserAchievements
      tags: [users]
      summary: Get user achievements
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/upcoming:
    get:
      operationId: listUpcomingEvents
      tags: [events]
      summary: List upcoming events
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List message board posts
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /merch:
    get:
      operationId: listMerch
      tags: [merch]
      summary: List merch products
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchProduct"

  /stats:
    get:
      operationId: getStats
      tags: [users]
      summary: Get community stats
      responses:
        "200":
          description: Community stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required:
        - email
        - password
        - name

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required:
        - user
        - token

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        isAdmin:
          type: boolean
        memberSince:
          type: string
        eventsAttended:
          type: integer
        medalsEarned:
          type: integer
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - email
        - name
        - isAdmin
        - memberSince
        - eventsAttended
        - medalsEarned

    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        eventId:
          type: integer
        attendedAt:
          type: string
        earnedMedal:
          type: boolean
        event:
          $ref: "#/components/schemas/Event"
      required:
        - id
        - userId
        - eventId
        - attendedAt
        - earnedMedal
        - event

    Achievement:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: string
          nullable: true
        unlocked:
          type: boolean
      required:
        - id
        - title
        - description
        - icon
        - unlocked

    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        location:
          type: string
        ticketUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isUpcoming:
          type: boolean
        attendeeCount:
          type: integer
      required:
        - id
        - title
        - description
        - date
        - location
        - isUpcoming
        - attendeeCount

    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        location:
          type: string
        ticketUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - title
        - description
        - date
        - location

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        authorName:
          type: string
        isMembersOnly:
          type: boolean
      required:
        - id
        - title
        - content
        - createdAt
        - authorName
        - isMembersOnly

    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        isMembersOnly:
          type: boolean
      required:
        - title
        - content
        - isMembersOnly

    MerchProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        buyUrl:
          type: string
          nullable: true
        category:
          type: string
        inStock:
          type: boolean
      required:
        - id
        - name
        - description
        - price
        - category
        - inStock

    CommunityStats:
      type: object
      properties:
        totalEvents:
          type: integer
        totalMembers:
          type: integer
        totalTicketsSold:
          type: integer
        totalMedalsAwarded:
          type: integer
      required:
        - totalEvents
        - totalMembers
        - totalTicketsSold
        - totalMedalsAwarded