# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dodge Club API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";

export interface AttendanceRecord {
  id: number;
  userId: number;
  eventId: number;
  attendedAt: string;
  earnedMedal: boolean;
  event: Event;
}