# lib/api-zod/src/generated/types/createPostRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dodge Club API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostRequest {
  title: string;
  content: string;
  imageUrl?: string | null;
  isMembersOnly: boolean;
}