# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dodge Club API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}