# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dodge Club API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  isAdmin: boolean;
  memberSince: string;
  eventsAttended: number;
  medalsEarned: number;
  avatarUrl?: string | null;
}